+++ src/main/java/com/cobaltplatform/api/integration/way2health/model/entity/Staff.java
@@ -21,7 +21,6 @@ package com.cobaltplatform.api.integration.way2health.model.entity;
 
 import javax.annotation.Nullable;
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigInteger;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -31,11 +30,11 @@ import java.util.List;
 @NotThreadSafe
 public class Staff extends Way2HealthEntity {
 	@Nullable
-	private BigInteger id;
+	private Long id;
 	@Nullable
-	private BigInteger accountId;
+	private Long accountId;
 	@Nullable
-	private BigInteger studyId;
+	private Long studyId;
 	@Nullable
 	private String name;
 	@Nullable
@@ -47,7 +46,7 @@ public class Staff extends Way2HealthEntity {
 	@Nullable
 	private String status;
 	@Nullable
-	private BigInteger notificationGroupId;
+	private Long notificationGroupId;
 	@Nullable
 	private String email;
 	@Nullable
@@ -68,16 +67,16 @@ public class Staff extends Way2HealthEntity {
 	@NotThreadSafe
 	public static class StudyGroup extends Way2HealthEntity {
 		@Nullable
-		private BigInteger id;
+		private Long id;
 		@Nullable
 		private String name;
 
 		@Nullable
-		public BigInteger getId() {
+		public Long getId() {
 			return id;
 		}
 
-		public void setId(@Nullable BigInteger id) {
+		public void setId(@Nullable Long id) {
 			this.id = id;
 		}
 
@@ -92,29 +91,29 @@ public class Staff extends Way2HealthEntity {
 	}
 
 	@Nullable
-	public BigInteger getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(@Nullable BigInteger id) {
+	public void setId(@Nullable Long id) {
 		this.id = id;
 	}
 
 	@Nullable
-	public BigInteger getAccountId() {
+	public Long getAccountId() {
 		return accountId;
 	}
 
-	public void setAccountId(@Nullable BigInteger accountId) {
+	public void setAccountId(@Nullable Long accountId) {
 		this.accountId = accountId;
 	}
 
 	@Nullable
-	public BigInteger getStudyId() {
+	public Long getStudyId() {
 		return studyId;
 	}
 
-	public void setStudyId(@Nullable BigInteger studyId) {
+	public void setStudyId(@Nullable Long studyId) {
 		this.studyId = studyId;
 	}
 
@@ -164,11 +163,11 @@ public class Staff extends Way2HealthEntity {
 	}
 
 	@Nullable
-	public BigInteger getNotificationGroupId() {
+	public Long getNotificationGroupId() {
 		return notificationGroupId;
 	}
 
-	public void setNotificationGroupId(@Nullable BigInteger notificationGroupId) {
+	public void setNotificationGroupId(@Nullable Long notificationGroupId) {
 		this.notificationGroupId = notificationGroupId;
 	}
