+++ src/main/java/com/cobaltplatform/api/integration/way2health/model/entity/TextMessage.java
@@ -21,7 +21,6 @@ package com.cobaltplatform.api.integration.way2health.model.entity;
 
 import javax.annotation.Nullable;
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -31,13 +30,13 @@ import java.util.List;
 @NotThreadSafe
 public class TextMessage extends Way2HealthEntity {
 	@Nullable
-	private BigInteger id;
+	private Long id;
 	@Nullable
 	private LocalDateTime createdAt;
 	@Nullable
 	private LocalDateTime updatedAt;
 	@Nullable
-	private BigInteger participantId;
+	private Long participantId;
 	@Nullable
 	private String messageText;
 	@Nullable
@@ -72,7 +71,7 @@ public class TextMessage extends Way2HealthEntity {
 	@NotThreadSafe
 	public static class Metadata extends Way2HealthEntity {
 		@Nullable
-		private BigInteger triggerId;
+		private Long triggerId;
 		@Nullable
 		private String message;
 		@Nullable
@@ -234,11 +233,11 @@ public class TextMessage extends Way2HealthEntity {
 	}
 
 	@Nullable
-	public BigInteger getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(@Nullable BigInteger id) {
+	public void setId(@Nullable Long id) {
 		this.id = id;
 	}
 
@@ -261,11 +260,11 @@ public class TextMessage extends Way2HealthEntity {
 	}
 
 	@Nullable
-	public BigInteger getParticipantId() {
+	public Long getParticipantId() {
 		return participantId;
 	}
 
-	public void setParticipantId(@Nullable BigInteger participantId) {
+	public void setParticipantId(@Nullable Long participantId) {
 		this.participantId = participantId;
 	}
